feat!: flip convention to qubits then qumodes
This flips the convention that for all hybrid gates, qubits come before
qumodes in the wire order, in order to better align with OpenQASM's
`ctrl @ gate` syntax. Many of the hybrid gates are conditioned on the
value of the qubit.

Because we're only in alpha version, we just bump to a new minor
version.